tools/vision/requirements.txt
Pillow>=10.0.0
opencv-python>=4.10.0
numpy>=1.26.0
